add new dao layer
