--- server/endpoints/auth/functions.ts
@@ -1,12 +1,0 @@
-import { OAuth2Client } from "google-auth-library";
-
-// eslint-disable-next-line import/prefer-default-export
-export const verifyTicket = async (token: string, audience: string) => {
-  const client = new OAuth2Client(audience);
-  const ticket = await client.verifyIdToken({
-    idToken: token,
-    audience,
-  });
-
-  return ticket.getPayload();
-};
