+++ server/src/api/admin/AdminChart.ts
@@ -1,8 +1,8 @@
 /* eslint-disable spaced-comment */
 import { body } from "express-validator";
-import { verifyToken } from "../utils/utils";
-import { Context, Endpoint } from "../../endpoints";
-import { Reviews, Classes, Subjects } from "../../db/dbDefs";
+import { verifyToken } from "../../utils/utils";
+import { Context, Endpoint } from "../../../endpoints";
+import { Reviews, Classes, Subjects } from "../../../db/dbDefs";
 import { GetReviewsOverTimeTop15Request, Token } from "./types";
 import { topSubjectsCB } from "./functions";
