+++ server/src/api/admin/functions.ts
@@ -1,7 +1,7 @@
-import { verifyToken } from "../utils/utils";
-import { Context } from "../../endpoints";
+import { verifyToken } from "../../utils/utils";
+import { Context } from "../../../endpoints";
 import { Token } from "./types";
-import { Reviews, Classes, Subjects } from "../../db/dbDefs";
+import { Reviews, Classes, Subjects } from "../../../db/dbDefs";
 import { DefaultDict } from "./AdminChart";
 
 /**
