+++ server/src/api/search/routes.ts
@@ -1,7 +1,7 @@
 import { body } from "express-validator";
 
-import { Context, Endpoint } from "../../endpoints";
-import { Classes, Subjects, Professors } from "../../db/dbDefs";
+import { Context, Endpoint } from "../../../endpoints";
+import { Classes, Subjects, Professors } from "../../../db/dbDefs";
 import { Search } from "./types";
 import { courseSort, regexClassesSearch } from "./functions";
