+++ server/src/api/search/types.ts
