+++ server/tsconfig.json
@@ -16,11 +16,11 @@
         "db/dbDefs.ts",
         "server.ts",
         "endpoints.ts",
-        "endpoints/admin/AdminActions.ts",
-        "endpoints/admin/AdminChart.ts",
-        "endpoints/auth/routes.ts",
-        "endpoints/review/routes.ts",
-        "endpoints/search/routes.ts",
-        "endpoints/utils/utils.ts"
+        "src/api/admin/AdminActions.ts",
+        "src/api/admin/AdminChart.ts",
+        "src/api/auth/routes.ts",
+        "src/api/review/routes.ts",
+        "src/api/search/routes.ts",
+        "src/utils/utils.ts"
     ]
 }
